Makes changes to displayMove() and corner handling
Gets rid of if/else in displayMove, previously it would
trigger a global corner variable, which is no longer
needed. Introduces a new corner play function
openingCorner() that will play the opposite corner
in the case of user opening with a corner. Rewrites
cornerPlay to not use global variable and limits the
if/else statments to if statments followed with returns.